+++ changelog.md
@@ -1,4 +1,8 @@
 ## General Changes
-- Added a playerutils class to handle player-related operations more efficiently & supported for folia.
+- Added reverse time parser (Ms -> String)
+- Added chance to custom inventory commands
+- Added new manager for log files
+- Plugin disable method can now be handled by zAPI#disable
+- Fixed & optimized find safe location function
 
-**Full Changelog**: https://github.com/yL3oft/zAPI/compare/1.4.4...1.4.5
+**Full Changelog**: https://github.com/yL3oft/zAPI/compare/1.4.5...1.4.6